Update gitignore, refactor reducer file name, add todo list
