+++ .gitignore
@@ -13,3 +13,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+#firebase
+/.firebase
