+++ src/container/ActiveUsersContainer.js
@@ -8,11 +8,13 @@ export class ActiveUsersContainer extends Component {
   render() {
     return (
       <div className="active-users__container">
-        active users
+        List of active users
       </div>
     )
   }
 }
+
+// TODO: Add a list of active users so user can select who they want to talk to
 
 // ActiveUsersContainer.propTypes = {
 //
