+++ src/reducers/rootReducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import {user, sendMessage, grabMessages} from './user';
+import {user, sendMessage, grabMessages} from './reducers';
 
 const rootReducer = combineReducers({
   user,
